Remove Quirk needsIPadMiniUserAgent for a series of domains https://bugs.webkit.org/show_bug.cgi?id=284103 rdar://140974340
Reviewed by NOBODY (OOPS!).

This list of domains had User Agent overrides to give an iPad Mini
UA instead of the usual desktop User Agent. This was done initially
to give most of the time UI compatible with touch events when the
website were unable to manage a device with a desktop UA and touch
interface. This list of domains has been tested and the UA override
is not required anymore.

* rdar://59480381 fidelity.com
* rdar://51885199 hsbc.com
* rdar://51355671 tving.com
* rdar://51355671 live.iqiyi.com
* rdar://51355671 video.sina.com.cn
* rdar://51355671 video.tudou.com
* rdar://51355671 v.china.com.cn
* rdar://49124313 trello.com
* rdar://51463649 ted.com
* rdar://67823088 nhl.com
* rdar://50839844 jsfiddle.net

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsIPadMiniUserAgent):